% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicTiles-class.R
\docType{methods}
\name{changeSettings}
\alias{changeSettings}
\alias{changeSettings,GenomicTiles,character-method}
\title{Check data compliance with tile settings}
\usage{
changeSettings(object, param, value)

\S4method{changeSettings}{GenomicTiles,character}(object, param, value)
}
\arguments{
\item{object}{A /code{GenomicTiles} object.}

\item{param}{The name of a tile settings parameter.}

\item{value}{An appropriate value. In most cases integer.}
}
\value{
A /code{GenomicTiles} object
}
\description{
Check if the indices were build correctly, according to the
specified parameters. This is the recommended way of changing
tile settings, as it triggers instant recomputation of the index.
}
\examples{
gt <- makeTestGenomicTiles()
gt2 <- changeSettings(gt, "chunkSize", 20)
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

