% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicTiles-class.R
\docType{methods}
\name{checkSettings}
\alias{checkSettings}
\alias{checkSettings,GenomicTiles-method}
\title{Check data compliance with tile settings}
\usage{
checkSettings(object)

\S4method{checkSettings}{GenomicTiles}(object)
}
\arguments{
\item{object}{A /code{GenomicTiles} object.}
}
\value{
A logical value
}
\description{
Check if the indices were build correctly, according to the
specified parameters
}
\examples{
gt <- makeTestGenomicTiles()
checkSettings(gt)
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

