% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvals.R
\name{computeSignificance}
\alias{computeSignificance}
\title{computeSignificance}
\usage{
computeSignificance(gg, log.p = FALSE)
}
\arguments{
\item{gg}{A GenoGAM object.}

\item{log.p}{Should values be returned in log scale?}
}
\value{
An updated GenoGAM object, where the pvalue slot is added.
}
\description{
The function computes positionwise p-values
}
\details{
Note, that in case the data is stored in HDF5 format, the pvalue 'group' is added on hard drive.
That is, unlike any other function in R, where the input object is not changed, it actually
is in this case. If one wishes to have HDF5 data without the pvalue 'group', one has to
backup the HDF5 files prior to computation or delete them after with \code{rhdf5::h5delete}
}
\examples{
## make test GenoGAM
gg <- makeTestGenoGAM()
## compute pvalues
computeSignificance(gg)
pvalue(gg)
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}
