% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{computeSizeFactors}
\alias{computeSizeFactors}
\title{computeSizeFactors}
\usage{
computeSizeFactors(ggd, factorGroups = NULL)
}
\arguments{
\item{ggd}{A GenoGAMDataSet object.}

\item{factorGroups}{A list of grouped IDs (same as the colnames of
the GenoGAMDataSet object). Each element of the list represents
a group of samples within which size factors are computed.
If NULL all samples are regarded to belong to one group.
Size factors are not computed between groups.}
}
\value{
An updated GenoGAMDataSet object.
}
\description{
The function computes the size factors for given factor groups based on
the DESeq2 package.
}
\examples{
ggd <- makeTestGenoGAMDataSet()
ggd <- computeSizeFactors(ggd)
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

