% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coordinates-class.R
\docType{class}
\name{Coordinates-class}
\alias{Coordinates-class}
\alias{Coordinates}
\alias{length,Coordinates-method}
\alias{nrow,Coordinates-method}
\alias{ncol,Coordinates-method}
\alias{start,Coordinates-method}
\alias{end,Coordinates-method}
\alias{width,Coordinates-method}
\alias{end<-,Coordinates-method}
\alias{start<-,Coordinates-method}
\title{Coordinates class}
\usage{
Coordinates(start = NULL, end = NULL)

\S4method{length}{Coordinates}(x)

\S4method{nrow}{Coordinates}(x)

\S4method{ncol}{Coordinates}(x)

\S4method{start}{Coordinates}(x)

\S4method{end}{Coordinates}(x)

\S4method{width}{Coordinates}(x)

\S4method{end}{Coordinates}(x) <- value

\S4method{start}{Coordinates}(x) <- value

\S4method{width}{Coordinates}(x)
}
\arguments{
\item{start}{The start vector, must be of same length as the end vector}

\item{end}{The end vector, must be of same length as the start vector}

\item{x}{The Coordinates object}

\item{value}{A numeric vector. For replace method in 'start' and 'end'.}
}
\value{
An object of class Coordinates
}
\description{
This class mainly exists to overcome the lack of long integer use in R
and the IRanges class in particular. The object is basically a restricted
DataFrame object, with a few custom methods to mimic the behaviour of IRanges.

Coordinates is the constructor function for the Coordinates-class.
}
\section{Methods (by generic)}{
\itemize{
\item \code{length}: The length method

\item \code{nrow}: The nrow method

\item \code{ncol}: The ncol method

\item \code{start}: The start accessor

\item \code{end}: The end accessor

\item \code{width}: The width accessor

\item \code{end<-}: Replacement method for end

\item \code{start<-}: Replacement method for end

\item \code{width}: The width accessor
}}

\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}
