% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicTiles-class.R
\docType{methods}
\name{dataRange}
\alias{dataRange}
\alias{dataRange,GPos-method}
\alias{dataRange,GenomicTiles-method}
\title{The /code{GRanges} of the underlying data}
\usage{
dataRange(object)

\S4method{dataRange}{GenomicTiles}(object)

\S4method{dataRange}{GPos}(object)
}
\arguments{
\item{object}{A /code{GenomicTiles} object.}
}
\value{
A /code{GRanges} object of genomic ranges of
the underlying data
}
\description{
Just like the /code{coordinates} slot but returns the genomic ranges
of the underlying data.
}
\examples{
gt <- makeTestGenomicTiles()
dataRange(gt)
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

