% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filterData}
\alias{filterData}
\title{A filter function for |code{GenoGAMDataSet}}
\usage{
filterData(ggd, threshold = NULL, windowsize = 201, mode = c("sum",
  "mean"))
}
\arguments{
\item{ggd}{A \code{GenoGAMDataSet} object}

\item{threshold}{A value for the mean or sum of counts, which will be used to filter on basepair level.
By default it is taken as median + 3*MAD}

\item{windowsize}{The sliding window size. Should be an odd value.}

\item{mode}{Should the sum or the mean of counts be used?}
}
\value{
A \code{GenoGAMDataSet} object containing the filtered regions
}
\description{
A function to filter the |code{GenoGAMDataSet} by the sum or mean of counts to significantly reduce
the amount of models to compute
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

