% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAM-class.R
\docType{class}
\name{GenoGAM-class}
\alias{GenoGAM-class}
\title{GenoGAM class}
\description{
This class is designed to represent the model object containing the estimate parameters,
arguments and finals fits of the model on a basepair level.
}
\section{Slots}{

\describe{
\item{\code{design}}{A mgcv-type formula object.}

\item{\code{fits}}{A data.frame of the fits, the standard error and the first and
second derivative of the fits for each experiment.}

\item{\code{positions}}{A GPos object of the positions and seqnames corresponding
to the rows in the 'fits' slot.}

\item{\code{smooths}}{A data.frame of knot positions and base function coefficients,
in order to reproduce the splines and compute derivatives.}

\item{\code{vcov}}{A list of covariance matrices for each tile fit.}

\item{\code{experimentDesign}}{The design matrix according to which the fitting
was performed.}

\item{\code{fitparams}}{Global parameters 'lambda', 'theta', 'Coefficient of Variation' and
the 'penalty order' used to compute the model.}

\item{\code{family}}{The distribution family.}

\item{\code{cvparams}}{Parameters used for cross validation.}

\item{\code{settings}}{The global and local settings that were used to compute the model.}

\item{\code{tileSettings}}{A list of settings used to compute tiles.}
}}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

