% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAM-class.R
\docType{methods}
\name{view,GenoGAM-method}
\alias{view,GenoGAM-method}
\title{View the dataset}
\usage{
\S4method{view}{GenoGAM}(object, ranges = NULL, seqnames = NULL,
  start = NULL, end = NULL)
}
\arguments{
\item{object}{A \code{GenoGAM} object}

\item{ranges}{A \code{GRanges} object. Makes it possible to
select regions by \code{GRanges}. Either ranges or seqnames, start and
end must be supplied}

\item{seqnames}{A chromosomes name. Either ranges or seqnames, start and
end must be supplied}

\item{start}{A start site. Either ranges or seqnames, start and
end must be supplied}

\item{end}{An end site. Either ranges or seqnames, start and
end must be supplied}
}
\value{
A data.frame of the selected data.
}
\description{
Cbinding the columns all together and coercing to data.frame
}
\examples{
gg <- makeTestGenoGAM()
gr <- GRanges("chrI", IRanges(1,40))
head(view(gg, gr))
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

