% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAM-package.R
\docType{package}
\name{GenoGAM}
\alias{GenoGAM}
\alias{GenoGAM-package}
\title{GenoGAM: A package providing a framework to analyse ChIP-Seq data}
\description{
This package allows statistical analysis of genome-wide
    data with smooth functions using generalized additive models
    based on the implementation from the R-package 'mgcv'. It
    provides methods for the statistical analysis of ChIP-Seq data
    including inference of protein occupancy, and pointwise and
    region-wise differential analysis. Estimation of dispersion and
    smoothing parameters is performed by cross-validation. Scaling
    of generalized additive model fitting to whole chromosomes is
    achieved by parallelization over overlapping genomic intervals.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gstricker/GenoGAM}
  \item Report bugs at \url{https://github.com/gstricker/GenoGAM/issues}
}

}
\author{
\strong{Maintainer}: Georg Stricker \email{georg.stricker@in.tum.de}

Authors:
\itemize{
  \item Alexander Engelhardt \email{alexander.engelhardt@ibe.med.uni-muenchen.de}
  \item Julien Gagneur \email{gagneur@in.tum.de}
}

}
