% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAMDataSet-class.R
\docType{methods}
\name{[,GenoGAMDataSet,GRanges,ANY,ANY-method}
\alias{[,GenoGAMDataSet,GRanges,ANY,ANY-method}
\title{Subsetting by GRanges}
\usage{
\S4method{[}{GenoGAMDataSet,GRanges,ANY,ANY}(x, i)
}
\arguments{
\item{x}{A \code{GenoGAMDataSet} object}

\item{i}{A \code{GRanges} object}
}
\value{
A subsetted \code{GenoGAMDataSet} object
}
\description{
Providing subsetting by GRanges through the single-bracket operator
}

