% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAMDataSet-class.R,
%   R/GenoGAMDataSetList-class.R
\docType{methods}
\name{Summary,GenoGAMDataSet-method}
\alias{Summary,GenoGAMDataSet-method}
\alias{mean,GenoGAMDataSet-method}
\alias{var,GenoGAMDataSet,ANY-method}
\alias{sd,GenoGAMDataSet-method}
\alias{median,GenoGAMDataSet-method}
\alias{mad,GenoGAMDataSet-method}
\alias{IQR,GenoGAMDataSet-method}
\alias{mean,GenoGAMDataSetList-method}
\alias{var,GenoGAMDataSetList,ANY-method}
\alias{sd,GenoGAMDataSetList-method}
\alias{median,GenoGAMDataSetList-method}
\alias{mad,GenoGAMDataSetList-method}
\alias{IQR,GenoGAMDataSetList-method}
\title{Computing metrics}
\usage{
\S4method{Summary}{GenoGAMDataSet}(x, ..., na.rm = FALSE)

\S4method{mean}{GenoGAMDataSet}(x)

\S4method{var}{GenoGAMDataSet,ANY}(x)

\S4method{sd}{GenoGAMDataSet}(x)

\S4method{median}{GenoGAMDataSet}(x)

\S4method{mad}{GenoGAMDataSet}(x)

\S4method{IQR}{GenoGAMDataSet}(x)

\S4method{mean}{GenoGAMDataSetList}(x)

\S4method{var}{GenoGAMDataSetList,ANY}(x)

\S4method{sd}{GenoGAMDataSetList}(x)

\S4method{median}{GenoGAMDataSetList}(x)

\S4method{mad}{GenoGAMDataSetList}(x)

\S4method{IQR}{GenoGAMDataSetList}(x)
}
\arguments{
\item{x}{A GenoGAMDataSet object}

\item{...}{Additional arguments}

\item{na.rm}{Should NAs be dropped. Otherwise the result is NA}
}
\value{
A matrix with the specified metric computed per tile per column
of the assay data.
}
\description{
Computing metrics on each tile of the GenoGAMDataSet object.
All metrics from the Summary generics group, as well as
mean, var, sd, median, mad and IQR are supported.
}
\examples{
ggd <- makeTestGenoGAMDataSet()
sum(ggd)
min(ggd)
max(ggd)
mean(ggd)
var(ggd)
sd(ggd)
median(ggd)
mad(ggd)
IQR(ggd)
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}
