% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAMDataSet-class.R
\docType{methods}
\name{subset,GenoGAMDataSet-method}
\alias{subset,GenoGAMDataSet-method}
\alias{subsetByOverlaps,GenoGAMDataSet,GRanges-method}
\alias{[,GenoGAMDataSet,GRanges,ANY,ANY-method}
\title{Subset methods for GenoGAMDataSet}
\usage{
\S4method{subset}{GenoGAMDataSet}(x, ...)

\S4method{subsetByOverlaps}{GenoGAMDataSet,GRanges}(x, ranges,
  maxgap = -1L, minoverlap = 0L, type = c("any", "start", "end",
  "within", "equal"), invert = FALSE, ...)

\S4method{[}{GenoGAMDataSet,GRanges,ANY,ANY}(x, i)
}
\arguments{
\item{x}{A GenoGAMDataSet object.}

\item{...}{Further arguments. Mostly a logical statement
in case of the 'subset' function. Note that the columnnames
for chromosomes and positions are: 'seqnames' and 'pos'.}

\item{ranges, i}{A GRanges object. In case of subsetting by double brackets
'i' is the index of the tile.}

\item{maxgap, minoverlap}{Intervals with a separation of 'maxgap' or
less and a minimum of 'minoverlap' overlapping positions, allowing for
'maxgap', are considered to be overlapping. 'maxgap' should
be a scalar, non-negative, integer. 'minoverlap' should be a scalar,
positive integer.}

\item{type}{By default, any overlap is accepted. By specifying the 'type'
parameter, one can select for specific types of overlap. The types correspond
to operations in Allen's Interval Algebra (see references in). If \code{type}
is \code{start} or \code{end}, the intervals are required to have matching
starts or ends, respectively. While this operation seems trivial, the naive
implementation using \code{outer} would be much less efficient. Specifying
\code{equal} as the type returns the intersection of the \code{start} and
\code{end} matches. If \code{type} is \code{within}, the query interval must
be wholly contained within the subject interval. Note that all matches must
additionally satisfy the \code{minoverlap} constraint described above.

The \code{maxgap} parameter has special meaning with the special
overlap types. For \code{start}, \code{end}, and \code{equal}, it specifies
the maximum difference in the starts, ends or both, respectively. For
\code{within}, it is the maximum amount by which the query may be wider
than the subject.}

\item{invert}{If TRUE, keep only the query ranges that do _not_ overlap
the subject.}
}
\value{
A subsetted GenoGAMDataSet object.
}
\description{
Subset methods for GenoGAMDataSet
}
\details{
Those are various methods to subset the GenoGAMDataSet object.
By logical statement or GRanges overlap. The '[' subsetter is
just a short version of 'subsetByOverlaps'.
}
\examples{

# subset by overlaps
ggd <- makeTestGenoGAMDataSet()
SummarizedExperiment::rowRanges(ggd)
gr <- GenomicRanges::GRanges("chrI", IRanges(10000,19000))
res <- IRanges::subsetByOverlaps(ggd, gr)
SummarizedExperiment::rowRanges(res)

# Subset by logical statement
ggd <- makeTestGenoGAMDataSet()
SummarizedExperiment::rowRanges(ggd)
res <- subset(ggd, seqnames == "chrI" & pos <= 17000)
SummarizedExperiment::rowRanges(res)
}
\references{
Allen's Interval Algebra: James F. Allen: Maintaining knowledge
about temporal intervals. In: Communications of the ACM.
26/11/1983. ACM Press. S. 832-843, ISSN 0001-0782
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}
