% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAMDataSetList-class.R
\docType{class}
\name{GenoGAMDataSetList-class}
\alias{GenoGAMDataSetList-class}
\alias{GenoGAMDataSetList}
\alias{dim,GenoGAMDataSetList-method}
\alias{length,GenoGAMDataSetList-method}
\alias{seqlengths,GenoGAMDataSetList-method}
\alias{seqlevels,GenoGAMDataSetList-method}
\alias{seqlevelsInUse,GenoGAMDataSetList-method}
\alias{colData,GenoGAMDataSetList-method}
\alias{rowRanges,GenoGAMDataSetList-method}
\alias{assay,GenoGAMDataSetList,ANY-method}
\alias{assays,GenoGAMDataSetList-method}
\alias{colnames,GenoGAMDataSetList-method}
\alias{getIndex,GenoGAMDataSetList-method}
\alias{getCountMatrix,GenoGAMDataSetList-method}
\alias{tileSettings,GenoGAMDataSetList-method}
\alias{dataRange,GenoGAMDataSetList-method}
\alias{getChromosomes,GenoGAMDataSetList-method}
\alias{getTileSize,GenoGAMDataSetList-method}
\alias{getChunkSize,GenoGAMDataSetList-method}
\alias{getOverhangSize,GenoGAMDataSetList-method}
\alias{getTileNumber,GenoGAMDataSetList-method}
\alias{is.HDF5,GenoGAMDataSetList-method}
\alias{design,GenoGAMDataSetList-method}
\alias{design<-,GenoGAMDataSetList,ANY-method}
\alias{sizeFactors,GenoGAMDataSetList-method}
\alias{sizeFactors<-,GenoGAMDataSetList,ANY-method}
\alias{getChunkSize<-,GenoGAMDataSetList,numeric-method}
\alias{getTileSize<-,GenoGAMDataSetList,numeric-method}
\alias{getOverhangSize<-,GenoGAMDataSetList,numeric-method}
\alias{getTileNumber<-,GenoGAMDataSetList,numeric-method}
\title{GenoGAMDataSetList}
\usage{
GenoGAMDataSetList(...)

\S4method{dim}{GenoGAMDataSetList}(x)

\S4method{length}{GenoGAMDataSetList}(x)

\S4method{seqlengths}{GenoGAMDataSetList}(x)

\S4method{seqlevels}{GenoGAMDataSetList}(x)

\S4method{seqlevelsInUse}{GenoGAMDataSetList}(x)

\S4method{colData}{GenoGAMDataSetList}(x, ...)

\S4method{rowRanges}{GenoGAMDataSetList}(x, ...)

\S4method{assay}{GenoGAMDataSetList,ANY}(x, i, ...)

\S4method{assays}{GenoGAMDataSetList}(x, ..., withDimnames = TRUE)

\S4method{colnames}{GenoGAMDataSetList}(x)

\S4method{getIndex}{GenoGAMDataSetList}(object)

\S4method{getCountMatrix}{GenoGAMDataSetList}(object)

\S4method{tileSettings}{GenoGAMDataSetList}(object)

\S4method{dataRange}{GenoGAMDataSetList}(object)

\S4method{getChromosomes}{GenoGAMDataSetList}(object)

\S4method{getTileSize}{GenoGAMDataSetList}(object)

\S4method{getChunkSize}{GenoGAMDataSetList}(object)

\S4method{getOverhangSize}{GenoGAMDataSetList}(object)

\S4method{getTileNumber}{GenoGAMDataSetList}(object)

\S4method{is.HDF5}{GenoGAMDataSetList}(object)

\S4method{design}{GenoGAMDataSetList}(object)

\S4method{design}{GenoGAMDataSetList,ANY}(object) <- value

\S4method{sizeFactors}{GenoGAMDataSetList}(object)

\S4method{sizeFactors}{GenoGAMDataSetList,ANY}(object) <- value

\S4method{getChunkSize}{GenoGAMDataSetList,numeric}(object) <- value

\S4method{getTileSize}{GenoGAMDataSetList,numeric}(object) <- value

\S4method{getOverhangSize}{GenoGAMDataSetList,numeric}(object) <- value

\S4method{getTileNumber}{GenoGAMDataSetList,numeric}(object) <- value
}
\arguments{
\item{...}{The slots and their respective values}

\item{withDimnames}{For use of S4 methods.}

\item{object, x}{For use of S4 methods. The GenoGAMDataSetList object.}

\item{value, i}{For use of S4 methods. The value to be assigned to the slot.}
}
\value{
An object of class GenoGAMDataSetList
}
\description{
The GenoGAMDataSetList class contains the pre-processed raw data and
additional slots that define the input and framework for the model.
It extends upon the idea of the GenoGAMDataSet class to make it possible
to store genomes and data of size > 2^31 (maximum size of integers in R).
Thus the only difference to a GenoGAMDataSet is the arrangement as a
list of RangedSummarizedExperiments under the hood. On the surface is
does still behave like a GenoGAMDataSet. It is not intended to be used
by the user. For more information check the GenoGAMDataSet class documentation.

GenoGAMDataSetList is the constructor function for the GenoGAMDataSetList-class.
}
\section{Methods (by generic)}{
\itemize{
\item \code{dim}: Get the dimension of the object

\item \code{length}: The length of the object

\item \code{seqlengths}: The seqlengths of the object

\item \code{seqlevels}: The seqlevels of the object

\item \code{seqlevelsInUse}: The seqlevelsInUse of the object

\item \code{colData}: get colData from the first element of the
SummarizedExperiment list

\item \code{rowRanges}: get a list of rowRanges from the
GenoGAMDataSetList object

\item \code{assay}: get a list of assays from the
GenoGAMDataSetList object

\item \code{assays}: get a list of list of assays from the
GenoGAMDataSetList object. Just for completeness, shouldn't be needed.

\item \code{colnames}: get colnames from the first element of the
SummarizedExperiment list

\item \code{getIndex}: accessor to the index slot

\item \code{getCountMatrix}: An accessor to the countMatrix slot

\item \code{tileSettings}: The accessor to the list of settings, that
were used to generate the tiles.

\item \code{dataRange}: The actual underlying GRanges showing the range of the data.

\item \code{getChromosomes}: A GRanges object representing the chromosomes
or chromosome regions on which the model will be computed

\item \code{getTileSize}: The size of the tiles

\item \code{getChunkSize}: The size of the chunks

\item \code{getOverhangSize}: The size of the overhang (on one side)

\item \code{getTileNumber}: The total number of tiles

\item \code{is.HDF5}: A boolean function that is true if object uses HDF5 backend

\item \code{design}: Access to the design slot.

\item \code{design<-}: Replace method of the design slot.

\item \code{sizeFactors}: Access to the sizeFactors slot

\item \code{sizeFactors<-}: Replace method of the sizeFactors slot

\item \code{getChunkSize<-}: Replace method of the chunkSize parameter,
that triggers a new computation of the tiles based on the new chunk size.

\item \code{getTileSize<-}: Replace method of the tileSize parameter,
that triggers a new computation of the tiles based on the new tile size.

\item \code{getOverhangSize<-}: Replace method of the overhangSize parameter,
that triggers a new computation of the tiles based on the new overhang size.

\item \code{getTileNumber<-}: Replace method of the tileNumber parameter,
that triggers a new computation of the tiles based on the new number of tiles.
}}

\section{Slots}{

\describe{
\item{\code{settings}}{The global and local settings that will be used to compute the
model.}

\item{\code{design}}{The formula describing how to evaluate the data. See details.}

\item{\code{sizeFactors}}{The normalized values for each sample. A named numeric vector.}

\item{\code{index}}{A GRanges object representing an index of the ranges defined 
on the genome. Mostly used to store tiles.}

\item{\code{data}}{A list of RangedSummarizedExperiment objects}

\item{\code{id}}{A GRanges object keeping the identifiers assigning the regions to the
respective list elements}

\item{\code{hdf5}}{A logical slot indicating if the object should be stored as HDF5}

\item{\code{countMatrix}}{Either a matrix or HDF5Matrix to store the sums of counts of
the regions (could also be seen as bins) for later use especially by DESeq2}
}}

\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}
