% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAMDataSetList-class.R
\docType{methods}
\name{Summary,GenoGAMDataSetList-method}
\alias{Summary,GenoGAMDataSetList-method}
\title{Computing metrics}
\usage{
\S4method{Summary}{GenoGAMDataSetList}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{A GenoGAMDataSetList object}

\item{...}{Additional arguments}

\item{na.rm}{Should NAs be dropped. Otherwise the result is NA}
}
\value{
A matrix with the specified metric computed per tile per column
of the assay data.
}
\description{
Computing metrics on each tile of the GenoGAMDataSetList object.
All metrics from the Summary generics group, as well as
mean, var, sd, median, mad and IQR are supported.
}
\examples{
ggd <- makeTestGenoGAMDataSetList()
sum(ggd)
min(ggd)
max(ggd)
mean(ggd)
var(ggd)
sd(ggd)
median(ggd)
mad(ggd)
IQR(ggd)
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}
