% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAMFamily-class.R
\docType{class}
\name{GenoGAMFamily-class}
\alias{GenoGAMFamily-class}
\title{GenoGAMFamily class}
\description{
This class holds the distribution family with the log-likelihood functions
and the first two derivatives, the gradient vector and the Hessian matrix.
It is not meant to be used directly, but rather for development convenience.
}
\section{Slots}{

\describe{
\item{\code{ll}}{The log-likelihood function. Gives a scalar}

\item{\code{gradient}}{The first derivative of the log-likelihood functions. Gives a vector.}

\item{\code{hessian}}{An integer indicating the family. Negative Binomial = 1, Quasi-Binomial = 2.}
}}

\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}
