% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAMList-class.R
\docType{class}
\name{GenoGAMList-class}
\alias{GenoGAMList-class}
\alias{GenoGAMList}
\alias{dim,GenoGAMList-method}
\alias{length,GenoGAMList-method}
\alias{seqlengths,GenoGAMList-method}
\alias{seqlevels,GenoGAMList-method}
\alias{seqlevelsInUse,GenoGAMList-method}
\alias{design,GenoGAMList-method}
\alias{sizeFactors,GenoGAMList-method}
\alias{getSettings,GenoGAMList-method}
\alias{getFamily,GenoGAMList-method}
\alias{colData,GenoGAMList-method}
\alias{getParams,GenoGAMList-method}
\alias{getCoefs,GenoGAMList-method}
\alias{getKnots,GenoGAMList-method}
\alias{assay,GenoGAMList,missing-method}
\alias{assays,GenoGAMList-method}
\alias{rowRanges,GenoGAMList-method}
\alias{fits,GenoGAMList-method}
\alias{se,GenoGAMList-method}
\alias{pvalue,GenoGAMList-method}
\alias{colnames,GenoGAMList-method}
\alias{dimnames,GenoGAMList-method}
\alias{is.HDF5,GenoGAMList-method}
\title{GenoGAMList class}
\usage{
GenoGAMList(..., ggd = NULL, fromHDF5 = FALSE)

\S4method{dim}{GenoGAMList}(x)

\S4method{length}{GenoGAMList}(x)

\S4method{seqlengths}{GenoGAMList}(x)

\S4method{seqlevels}{GenoGAMList}(x)

\S4method{seqlevelsInUse}{GenoGAMList}(x)

\S4method{design}{GenoGAMList}(object)

\S4method{sizeFactors}{GenoGAMList}(object)

\S4method{getSettings}{GenoGAMList}(object)

\S4method{getFamily}{GenoGAMList}(object)

\S4method{colData}{GenoGAMList}(x)

\S4method{getParams}{GenoGAMList}(object)

\S4method{getCoefs}{GenoGAMList}(object)

\S4method{getKnots}{GenoGAMList}(object)

\S4method{assay}{GenoGAMList,missing}(x, i)

\S4method{assays}{GenoGAMList}(x, ..., withDimnames = TRUE)

\S4method{rowRanges}{GenoGAMList}(x, ...)

\S4method{fits}{GenoGAMList}(object)

\S4method{se}{GenoGAMList}(object)

\S4method{pvalue}{GenoGAMList}(object)

\S4method{colnames}{GenoGAMList}(x)

\S4method{dimnames}{GenoGAMList}(x)

\S4method{is.HDF5}{GenoGAMList}(object)
}
\arguments{
\item{...}{Slots of the GenoGAM class. See the slot description.}

\item{ggd}{The initial GenoGAMDataSet object. Only needed if fromHDF5 is TRUE.}

\item{fromHDF5}{A convenience argument to create a GenoGAM object from the already
computed fits that are stored as HDF5 files}

\item{object, x}{For use of S4 methods. The GenoGAMList object.}

\item{i}{A GRanges object (only for subsetting)}

\item{withDimnames}{For use of S4 methods. The GenoGAMList object.}
}
\value{
An object of the type GenoGAM.
}
\description{
This is the class the holds the complete model as well all hyperparameters
and settings that were used to fit it. It is basically identical to the
GenoGAM class, except for the data being inside a list of RangedSummarizedExperiment
objects.

The GenoGAMList constructor, not designed to be actually used, by the user.
Rather to be a point of reference and documentation for slots and how
to access them.
}
\section{Methods (by generic)}{
\itemize{
\item \code{dim}: Get the dimension of the object

\item \code{length}: The length of the object

\item \code{seqlengths}: The seqlengths of the object

\item \code{seqlevels}: The seqlevels of the object

\item \code{seqlevelsInUse}: The seqlevelsInUse of the object

\item \code{design}: An accessor to the design slot

\item \code{sizeFactors}: An accessor to the sizeFactors slot

\item \code{getSettings}: An accessor to the settings slot

\item \code{getFamily}: An accessor to the family slot

\item \code{colData}: An accessor to the factorialDesign slot.

\item \code{getParams}: An accessor to the params slot

\item \code{getCoefs}: An accessor to the coefs slot

\item \code{getKnots}: An accessor to the knots slot

\item \code{assay}: The accessor to the fits and standard errors

\item \code{assays}: get a list of list of assays from the
GenoGAMList object. Just for completeness, shouldn't be needed.

\item \code{rowRanges}: get a list of rowRanges from the
GenoGAMList object

\item \code{fits}: An accessor to the fits

\item \code{se}: An accessor to the standard errors

\item \code{pvalue}: An accessor to the pvalues

\item \code{colnames}: column names of GenoGAMList

\item \code{dimnames}: The names of the dimensions of GenoGAMList

\item \code{is.HDF5}: A boolean function that is true if object uses HDF5 backend
}}

\section{Slots}{

\describe{
\item{\code{family}}{The name of the distribution family used}

\item{\code{design}}{The formula of the model}

\item{\code{sizeFactors}}{The offset used in the model.}

\item{\code{factorialDesign}}{The factorial design used. The same as colData in the
GenoGAMDataSet}

\item{\code{params}}{All hyperparameters used to fit the data. The parameters
estimated by cross validation can also be found here. But the parameters
used in cross validation are in the settings slot.}

\item{\code{settings}}{A GenoGAMSettings object representing the global
settings that were used to compute the model.}

\item{\code{data}}{A list of RangedSummarizedExperiment that holds the actual data}

\item{\code{id}}{A GRanges object keeping the identifiers assigning the regions to the
respective list elements}

\item{\code{coefs}}{The coefficients of the knots}

\item{\code{knots}}{The relative knot positions}
}}

\examples{

## creating test GenoGAM object
gg <- makeTestGenoGAM()
gg

## using accessors
design(gg)
sizeFactors(gg)
getSettings(gg)
getFamily(gg)
colData(gg)
getParams(gg)
getCoefs(gg)
getKnots(gg)
rowRanges(gg)
assay(gg)
assays(gg)
fits(gg) 
se(gg)
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}
