% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAMSettings-class.R
\name{GenoGAMSettings}
\alias{GenoGAMSettings}
\title{The constructor function for GenoGAMSettings}
\usage{
GenoGAMSettings(...)
}
\arguments{
\item{...}{Any parameters corresponding to the slots and their possible
values. See \linkS4class{GenoGAMSettings}}
}
\value{
A GenoGAMSettings object.
}
\description{
The constructor function for GenoGAMSettings
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

