% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicTiles-class.R
\docType{methods}
\name{[[,GenomicTiles,numeric,ANY-method}
\alias{[,GenomicTiles,GRanges,ANY,ANY-method}
\alias{[[,GenomicTiles,numeric,ANY-method}
\title{Providing pseudo-list functionality}
\usage{
\S4method{[[}{GenomicTiles,numeric,ANY}(x, i)

\S4method{[}{GenomicTiles,GRanges,ANY,ANY}(x, i)
}
\arguments{
\item{x}{A \code{GenomicTiles} object}

\item{i}{An integer (for '[[') or a \code{GRanges} object (for '[')}
}
\value{
A \code{DataFrame} (for '[[') or a subsetted \code{GenomicTiles} object (for '[')
}
\description{
Getting a specific tile
}

