% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicTiles-class.R
\docType{class}
\name{GenomicTiles-class}
\alias{GenomicTiles-class}
\title{GenomicTiles class}
\description{
This class is designed to represent the entire genome (or a subset of
it) and any additional data associated with the samples or positions.
It extends the RangedSummarizedExperiment class and adds two additional
index slots to keep track of the data. The main change compared to
RangedSummarizedExperiment is the use of a GPos (basepair level) instead of
GRanges (ranges level) object as rowRanges and the use of two GRanges objects
as indices. The GPos object allows to store raw instead of summarized data
in the assays. Because of this the size of genomic data can increase
tremendously. Thus the GenomicTiles class automatically divides the data
in (overlapping) tiles, making any operation on this data easy executable
in parallel.
}
\details{
For all other slots see \code{\link[SummarizedExperiment]{SummarizedExperiment}}.
}
\section{Slots}{

\describe{
\item{\code{index}}{A GRanges object that stores the tiles ranges and their index
in the genome space. That is, ranges are the positions on the genome.}

\item{\code{coordinates}}{A GRanges object that stores the tiles ranges and their index
in the DataFrame space. That is ranges are the row positions in the DataFrame.}
}}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

