% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicTiles-class.R
\docType{methods}
\name{Summary,GenomicTiles-method}
\alias{IQR,GenomicTiles-method}
\alias{Summary,GenomicTiles-method}
\alias{mad,GenomicTiles-method}
\alias{mean,GenomicTiles-method}
\alias{median,GenomicTiles-method}
\alias{sd,GenomicTiles-method}
\alias{var,GenomicTiles,ANY-method}
\title{Computing metrics}
\usage{
\S4method{Summary}{GenomicTiles}(x, ..., na.rm = FALSE)

\S4method{mean}{GenomicTiles}(x)

\S4method{var}{GenomicTiles,ANY}(x)

\S4method{sd}{GenomicTiles}(x)

\S4method{median}{GenomicTiles}(x)

\S4method{mad}{GenomicTiles}(x)

\S4method{IQR}{GenomicTiles}(x)
}
\arguments{
\item{x}{A \code{GenomicTiles} object}

\item{...}{Additional arguments}

\item{na.rm}{Should NAs be dropped. Otherwise the result is NA}
}
\value{
A list of as many elements as there are assays.
Each element contains of a matrix with the specified
metric computed per tile per column of the assay data.
}
\description{
Computing metrics on each tile of the \code{GenomicTiles} object.
So far all metrics from the Summary generics group, as well as
mean, var, sd, median, mad and IQR are supported.
}
\examples{
gt <- makeTestGenomicTiles()
sum(gt)
min(gt)
max(gt)
mean(gt)
var(gt)
sd(gt)
median(gt)
mad(gt)
IQR(gt)
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

