% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicTiles-class.R
\docType{methods}
\name{view}
\alias{view}
\alias{view,GenomicTiles-method}
\title{View the dataset}
\usage{
view(object, ...)

\S4method{view}{GenomicTiles}(object, ranges = NULL, seqnames = NULL,
  start = NULL, end = NULL)
}
\arguments{
\item{object}{A \code{GenomicTiles} object}

\item{...}{Additional arguments}

\item{ranges}{A \code{GRanges} object. Makes it possible to
select regions by \code{GRanges}. Either ranges or seqnames, start and
end must be supplied}

\item{seqnames}{A chromosomes name. Either ranges or seqnames, start and
end must be supplied}

\item{start}{A start site. Either ranges or seqnames, start and
end must be supplied}

\item{end}{An end site. Either ranges or seqnames, start and
end must be supplied}
}
\value{
A data.frame of the selected data.
}
\description{
Cbinding the columns all together and coercing to data.frame
}
\examples{
gt <- makeTestGenomicTiles()
gr <- GRanges(c("chrI", "chrII"), IRanges(c(1, 10), c(40, 30)))
head(view(gt, ranges = gr))
head(view(gt, seqnames = "chrI", start = 1, end = 20))
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

