% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicTiles-class.R
\name{GenomicTiles}
\alias{GenomicTiles}
\title{GenomicTiles constructor.}
\usage{
GenomicTiles(assays, chunkSize = 10000, overhangSize = 0, ...)
}
\arguments{
\item{assays}{One of two things. Either directly an object of
type 'RangedSummarizedExperiment'. Or in case the object is
created from raw data, a 'list' or 'SimpleList' of matrix-like
elements, or a matrix-like object. All elements of the list must
have the same dimensions, and dimension names (if present) must
be consistent across elements and with the row names of 'rowRanges'
and 'colData'.}

\item{chunkSize}{An integer specifying the size of one chunk in bp.}

\item{overhangSize}{An integer specifying the size of the overhang in bp.
The overhang is regarded to be symmetric, such that only the overhang
of one side should be provided.}

\item{...}{Further parameters passed to the
\code{\link[SummarizedExperiment]{SummarizedExperiment}} constructor.}
}
\value{
An object of class GenomicTiles.
}
\description{
This is the constructor function for GenomicTiles. The easiest
construction is fropm SummarizedExperiment. However as the
class operates on basepair level, the rowRanges are restricted
to the GPos class.
}
\details{
Most, but not necessary all functionalities of
SummarizedExperiment are yet provided.
}
\examples{
## from raw data
gp <- GPos(GRanges(c("chrI", "chrII"), IRanges(c(1,1), c(5,5))))
assay <- matrix(1:10, 10, 1)
gt <- GenomicTiles(assay, chunkSize = 3, rowRanges = gp)

## from SummarizedExperiment
se <- SummarizedExperiment(assay, rowRanges = gp)
gt <- GenomicTiles(se, chunkSize = 3)
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

