% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicTiles-class.R
\docType{methods}
\name{getChunkIndex}
\alias{getChunkIndex}
\alias{getChunkIndex,GenomicTiles-method}
\title{Compute the index for chunks instead tiles}
\usage{
getChunkIndex(object, ...)

\S4method{getChunkIndex}{GenomicTiles}(object, id = NULL)
}
\arguments{
\item{object}{A /code{GenomicTiles} object.}

\item{...}{Additional arguments}

\item{id}{A vector if tile ids. By default the complete index is returned.}
}
\value{
A /code{GRanges} object representing the index
}
\description{
The chunk index holds the Granges object that splits the entire dataset in
chunk, that is non-overlapping intervals.
}
\examples{
gt <- makeTestGenomicTiles()
getChunkIndex(gt)
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

