% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicTiles-class.R
\docType{methods}
\name{getCoordinates}
\alias{getCoordinates}
\alias{getCoordinates,GenomicTiles-method}
\title{Accessor to the /code{coordinates} slot}
\usage{
getCoordinates(object)

\S4method{getCoordinates}{GenomicTiles}(object)
}
\arguments{
\item{object}{A /code{GenomicTiles} object.}
}
\value{
A /code{GRanges} object of row coordinates
}
\description{
The /code{coordinates} slot contains the row coordinates of each
chromosome in the data. Such that taken a genomic position from
a chromosome it's easy to detect the correct row in the assay
}
\examples{
gt <- makeTestGenomicTiles()
getCoordinates(gt)
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

