% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAM-class.R
\docType{methods}
\name{getFits}
\alias{getFits}
\alias{getFits,GenoGAM-method}
\title{Accessor to \code{fits} slot}
\usage{
getFits(object)

\S4method{getFits}{GenoGAM}(object)
}
\arguments{
\item{object}{A \code{GenomicTiles} object}
}
\value{
A data.frame of the fits
}
\description{
The \code{fits} slot contains the fitted values of the model
}
\examples{
gg <- makeTestGenoGAM()
fits <- getFits(gg)
}

