% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicTiles-class.R
\docType{methods}
\name{getIndex}
\alias{getIndex}
\alias{getIndex,GenomicTiles-method}
\title{Accessor to the 'index' slot}
\usage{
getIndex(object, ...)

\S4method{getIndex}{GenomicTiles}(object, id = NULL)
}
\arguments{
\item{object}{A /code{GenomicTiles} object.}

\item{...}{Additional arguments}

\item{id}{A vector if tile ids. By default the complete index is returned.}
}
\value{
A \code{GRanges} object representing the index
}
\description{
The index holds the Granges object that splits the entire dataset in
tiles.
}
\examples{
gt <- makeTestGenomicTiles()
getIndex(gt)
getIndex(gt, 1:3)
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

