% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicTiles-class.R
\docType{methods}
\name{getIndexCoordinates}
\alias{getIndexCoordinates}
\alias{getIndexCoordinates,GenomicTiles-method}
\title{Compute the row coordinates for a given index}
\usage{
getIndexCoordinates(object, ...)

\S4method{getIndexCoordinates}{GenomicTiles}(object, id = NULL,
  index = NULL)
}
\arguments{
\item{object}{A /code{GenomicTiles} object.}

\item{...}{Additional arguments
Usually the original index or the chunk index.}

\item{id}{A vector if tile ids. By default the complete index is returned.}

\item{index}{A /code{Granges} object representing an index of genomic positions.}
}
\value{
A /code{GRanges} object of row coordinates
}
\description{
Given an index of genomic positions, this method computes
the corresponding row positions in the assay
}
\examples{
gt <- makeTestGenomicTiles()
getIndexCoordinates(gt)
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

