% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicTiles-class.R
\docType{methods}
\name{getTile}
\alias{getTile}
\alias{getTile,GenomicTiles-method}
\title{Tile extraction as a DataFrame}
\usage{
getTile(object, id, ...)

\S4method{getTile}{GenomicTiles}(object, id, size = 3e+09)
}
\arguments{
\item{object}{A \code{GenomicTiles} object}

\item{id}{A vector of tile ids}

\item{...}{Additional arguments}

\item{size}{The maximal number of rows that should be handled at once.
If the dataset is bigger it will be processed in chunks. This is to lower
memory consumption on big datasets, which in turn is slower.}
}
\value{
A \code{SimpleDataFrameList}
}
\description{
Extracting one or multiple tiles from a \code{GenomicTiles} object
and coercing them to a DataFrameList.
}
\examples{
gt <- makeTestGenomicTiles()
getTile(gt, 1:3)
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

