% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAM-class.R
\name{makeTestGenoGAM}
\alias{makeTestGenoGAM}
\title{Make an example /code{GenoGAM}}
\usage{
makeTestGenoGAM()
}
\value{
A /code{GenoGAM} object
}
\description{
Make an example /code{GenoGAM}
}
\examples{
gg <- makeTestGenoGAM()
}
