% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAMDataSet-class.R
\name{makeTestGenoGAMDataSet}
\alias{makeTestGenoGAMDataSet}
\title{Make an example /code{GenoGAMDataSet}}
\usage{
makeTestGenoGAMDataSet(sim = FALSE)
}
\arguments{
\item{sim}{Use simulated data (TRUE) or test data from a real experiment}
}
\value{
A /code{GenoGAMDataSet} object
}
\description{
Make an example /code{GenoGAMDataSet}
}
\examples{
realdt <- makeTestGenoGAMDataSet()
simdt <- makeTestGenoGAMDataSet(sim = TRUE)
}
