% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAMDataSetList-class.R
\name{makeTestGenoGAMDataSetList}
\alias{makeTestGenoGAMDataSetList}
\title{Make an example /code{GenoGAMDataSet}}
\usage{
makeTestGenoGAMDataSetList()
}
\value{
A /code{GenoGAMDataSet} object
}
\description{
Make an example /code{GenoGAMDataSet}
}
\examples{
ggdl <- makeTestGenoGAMDataSetList()
}
