% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAMList-class.R
\name{makeTestGenoGAMList}
\alias{makeTestGenoGAMList}
\title{Make an example /code{GenoGAMList}}
\usage{
makeTestGenoGAMList()
}
\value{
A /code{GenoGAMList} object
}
\description{
Make an example /code{GenoGAMList}
}
\examples{
ggl <- makeTestGenoGAMList()
}
