% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicTiles-class.R
\name{makeTestGenomicTiles}
\alias{makeTestGenomicTiles}
\title{Make an example /code{GenomicTile}}
\usage{
makeTestGenomicTiles()
}
\value{
A /code{GenomicTiles} object
}
\description{
Make an example /code{GenomicTile}
}
\examples{
test <- makeTestGenomicTiles()
}

