% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.GenoGAM}
\alias{plot.GenoGAM}
\title{The pot function for a GenoGAM object}
\usage{
plot.GenoGAM(x, ggd = NULL, ranges = NULL, seqnames = NULL,
  start = NULL, end = NULL, scale = TRUE, ...)
}
\arguments{
\item{x}{A \code{GenoGAM} object}

\item{ggd}{A \code{GenoGAMDataSet} object to plot raw counts}

\item{ranges}{A \code{GRanges} object specifying a particular region}

\item{seqnames}{A chromosome name. Together with start and end it is an
alternative way ob selecting a region}

\item{start}{The start of a region}

\item{end}{The end of a region}

\item{scale}{Logical, should all tracks be scaled to the same y-axis?}

\item{...}{Additional parameters that will be passed to the basic plot routine}
}
\value{
A plot of all tracks either using the ggplot2 or the base R framework
}
\description{
This functions plots the fit of a given region and optionally the read counts from the GenoGAMDataSet object
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

