% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAM-class.R
\docType{methods}
\name{subset,GenoGAM-method}
\alias{subset,GenoGAM-method}
\title{Subset method for \code{GenoGAM}}
\usage{
\S4method{subset}{GenoGAM}(x, ...)
}
\arguments{
\item{x}{A \code{GenoGAM} object.}

\item{...}{Further arguments. Mostly a logical statement.
Note that the columnnames for chromosomes and positions
are: seqnames and pos.}
}
\value{
A subsetted \code{GenoGAM} object.
}
\description{
Subsetting the \code{GenoGAM} by a logical statement
}
\examples{
gg <- makeTestGenoGAM()
subset(gg, pos <= 40)
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

