% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAMDataSet-class.R
\docType{methods}
\name{subset,GenoGAMDataSet-method}
\alias{subset,GenoGAMDataSet-method}
\title{Subset method for \code{GenoGAMDataSet}}
\usage{
\S4method{subset}{GenoGAMDataSet}(x, ...)
}
\arguments{
\item{x}{A \code{GenoGAMDataSet} object.}

\item{...}{Further arguments. Mostly a logical statement.
Note that the columnnames for chromosomes and positions
are: seqnames and pos.}
}
\value{
A subsetted \code{GenomicTiles} object.
}
\description{
Subsetting the \code{GenoGAMDataSet} by a logical statement
}
\examples{
ggd <- makeTestGenoGAMDataSet()
res <- subset(ggd, seqnames == "chrI" & pos <= 50)
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

