% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicTiles-class.R
\docType{methods}
\name{subset,GenomicTiles-method}
\alias{subset,GenomicTiles-method}
\title{Subset method for /code{GenomciTiles}}
\usage{
\S4method{subset}{GenomicTiles}(x, ...)
}
\arguments{
\item{x}{A /code{GenomicTiles} object.}

\item{...}{Further arguments. Mostly a logical statement.
Note that the columnnames for chromosomes and positions
are: seqnames and pos.}
}
\value{
A subsetted /code{GenomicTiles} object.
}
\description{
Subsetting the /code{GenomicTiles} by a logical statement
}
\examples{
gt <- makeTestGenomicTiles()
res <- subset(gt, seqnames == "chrI" & pos <= 50)
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

