% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAMDataSet-class.R
\docType{methods}
\name{subsetByOverlaps,GenoGAMDataSet,GRanges-method}
\alias{subsetByOverlaps,GenoGAMDataSet,GRanges-method}
\title{Subset by overlaps method for \code{GenoGAMDataSet}}
\usage{
\S4method{subsetByOverlaps}{GenoGAMDataSet,GRanges}(query, subject,
  maxgap = 0L, minoverlap = 1L, type = c("any", "start", "end", "within",
  "equal"), ...)
}
\arguments{
\item{query}{A \code{GenoGAMDataSet} object.}

\item{subject}{A \code{GRanges} object}

\item{maxgap, minoverlap}{Intervals with a separation of \code{maxgap} or
less and a minimum of \code{minoverlap} overlapping positions, allowing for
\code{maxgap}, are considered to be overlapping.  \code{maxgap} should
be a scalar, non-negative, integer. \code{minoverlap} should be a scalar,
positive integer.}

\item{type}{By default, any overlap is accepted. By specifying the \code{type}
parameter, one can select for specific types of overlap. The types correspond
to operations in Allen's Interval Algebra (see references). If \code{type}
is \code{start} or \code{end}, the intervals are required to have matching
starts or ends, respectively. While this operation seems trivial, the naive
implementation using \code{outer} would be much less efficient. Specifying
\code{equal} as the type returns the intersection of the \code{start} and
\code{end} matches. If \code{type} is \code{within}, the query interval must
be wholly contained within the subject interval. Note that all matches must
additionally satisfy the \code{minoverlap} constraint described above.

The \code{maxgap} parameter has special meaning with the special
overlap types. For \code{start}, \code{end}, and \code{equal}, it specifies
the maximum difference in the starts, ends or both, respectively. For
\code{within}, it is the maximum amount by which the query may be wider
than the subject.}

\item{...}{Additional parameters}
}
\value{
A subsetted \code{GenoGAMDataSet} object.
}
\description{
Subsetting the \code{GenoGAMDataSet} by a \code{GRanges} object
}
\examples{
ggd <- makeTestGenoGAMDataSet()
gr <- GRanges("chrI", IRanges(1,50))
res <- subsetByOverlaps(ggd, gr)
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

