% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicTiles-class.R
\docType{methods}
\name{tileSettings}
\alias{tileSettings}
\alias{tileSettings,GenomicTiles-method}
\title{Return tile settings}
\usage{
tileSettings(object)

\S4method{tileSettings}{GenomicTiles}(object)
}
\arguments{
\item{object}{A /code{GenomicTiles} object.}
}
\value{
A list of tile settings
}
\description{
Returns a list settings used to generate the tile index
}
\examples{
gt <- makeTestGenomicTiles()
tileSettings(gt)
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

