% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicTiles-class.R
\docType{methods}
\name{untile}
\alias{untile}
\alias{untile,GenomicTiles-method}
\title{Set index to chunkIndex}
\usage{
untile(object, ...)

\S4method{untile}{GenomicTiles}(object, id = NULL)
}
\arguments{
\item{object}{A /code{GenomicTiles} object.}

\item{...}{Additional arguments}

\item{id}{A vector if tile ids. By default the complete index is taken.}
}
\value{
A modified /code{GenomicTiles} object
}
\description{
Replace the tile index with the chunk index in /code{GenomicTiles} object
}
\examples{
gt <- makeTestGenomicTiles()
newGT <- untile(gt)
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

