% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffBinding.R
\name{computeRegionSignificance}
\alias{computeRegionSignificance}
\title{Compute significance for given regions}
\usage{
computeRegionSignificance(fit, regions, what)
}
\arguments{
\item{fit}{A GenoGAM object containing the fit}

\item{regions}{A GRanges object of regions of interest}

\item{what}{Which fit should be used. The names should be equivalent to the column names used in the config file.
Lookup with \code{names(colData(my_GenoGAMDataSet_object))}}
}
\value{
The GRanges object from the 'region' parameter extended by two columns: pvalue and FDR
}
\description{
For a given set of regions, region-wise pvalues and FDR is computed
}
\details{
For a given set of regions, region-wise pvalues are computed by applying
familywise hochberg correction and taking the minimal p-value. FDR is computed by further
applying Benjamini-Hochberg correction.
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

