% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAM-class.R
\name{computeSignificance}
\alias{computeSignificance}
\title{Compute significance.}
\usage{
computeSignificance(gg, log.p = FALSE)
}
\arguments{
\item{gg}{A fitted GenoGAM object.}

\item{log.p}{Should pvalues be returned in log scale?}
}
\value{
A GenoGAM object which fits has been updated by the pvalue columns.
}
\description{
Based on the model fits this functions computes pointwise pvalues.
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

