% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAM-class.R
\docType{methods}
\name{GenoGAM-methods}
\alias{GenoGAM-methods}
\alias{colData,GenoGAM-method}
\alias{design,GenoGAM-method}
\alias{getFits}
\alias{getFits,GenoGAM-method}
\alias{rowRanges,GenoGAM-method}
\title{GenoGAM-methods}
\usage{
\S4method{rowRanges}{GenoGAM}(x)

\S4method{design}{GenoGAM}(object)

getFits(x)

\S4method{getFits}{GenoGAM}(x)

\S4method{colData}{GenoGAM}(x)
}
\arguments{
\item{x, object}{A GenoGAM object.}
}
\value{
The respective slot
}
\description{
The different accessor functions for the GenoGAM object

The 'positions' slot holds the positions of the fit in GPos format

The 'design' slot holds the formula of the fit

The 'fits' slot contains the fitted values of the model

The 'experimentDesign' slot contains the experimental design of the model as
specified in the config file
}
\examples{
gg <- makeTestGenoGAM()
ranges <- rowRanges(gg)
gg <- makeTestGenoGAM()
des <- design(gg)
gg <- makeTestGenoGAM()
fits <- getFits(gg)
gg <- makeTestGenoGAM()
exdesign <- colData(gg)
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

