% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAM-class.R
\docType{methods}
\name{rowRanges,GenoGAM-method}
\alias{colData,GenoGAM-method}
\alias{design,GenoGAM-method}
\alias{getFits}
\alias{getFits,GenoGAM-method}
\alias{rowRanges,GenoGAM-method}
\title{Accessor to the 'positions' slot}
\usage{
\S4method{rowRanges}{GenoGAM}(x)

\S4method{design}{GenoGAM}(object)

getFits(object)

\S4method{getFits}{GenoGAM}(object)

\S4method{colData}{GenoGAM}(x)
}
\arguments{
\item{object}{A GenoGAM object.}

\item{object}{A formula object.}

\item{object}{A GenoGAM object}

\item{object}{A GenoGAM object}
}
\value{
A GPos object representing the positions

The formula of the fit

A data.frame of the fits

A named matrix
}
\description{
The positions slot holds the positions of the fits as a |code{GPos} object

The 'design' slot holds the formula of the fit

The 'fits' slot contains the fitted values of the model

The 'experimentDesign' slot contains the experimental design of the model as
specified in the config file
}
\examples{
gg <- makeTestGenoGAM()
fits <- getFits(gg)
gg <- makeTestGenoGAM()
fits <- colData(gg)
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}

Georg Stricker \email{georg.stricker@in.tum.de}
}

