% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAM-class.R
\docType{methods}
\name{rowRanges,GenoGAM-method}
\alias{rowRanges,GenoGAM-method}
\title{Accessor to the 'positions' slot}
\usage{
\S4method{rowRanges}{GenoGAM}(x)
}
\arguments{
\item{object}{A /code{GenoGAM} object.}
}
\value{
A |code{GPos} object representing the positions
}
\description{
The positions slot holds the positions of the fits as a |code{GPos} object
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

