% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAMDataSet-class.R
\docType{class}
\name{GenoGAMDataSet-class}
\alias{GenoGAMDataSet-class}
\title{GenoGAMDataSet}
\description{
This class is designed to represent the input for the GenoGAM model.
it extends the GenomicTiles class.
}
\details{
For all other slots see SummarizedExperiment.
}
\section{Slots}{

\describe{
\item{\code{settings}}{The global and local settings that were used to compute the model.}

\item{\code{design}}{The formula describing how to evaluate the data.}

\item{\code{sizeFactors}}{The normalized values for each sample.}
}}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

