% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAMSettings-class.R
\docType{class}
\name{GenoGAMSettings-class}
\alias{GenoGAMSettings-class}
\title{GenoGAMSettings}
\description{
This class is designed to store settings for the computation of the
GenoGAM package
}
\details{
Center can have three values: TRUE, FALSE, NULL. TRUE will
trigger the center function, FALSE will trigger the use of the entire
fragment. NULL should be used in case a custom process function is used.
}
\section{Slots}{

\describe{
\item{\code{center}}{A logical or NULL value to specify if the raw data should be
centered, i.e. only the midpoint of the fragment will be used to
represent its coverage. See details.}

\item{\code{chromosomeList}}{A character vector of chromosomes to be used.
NULL for all chromosomes.}

\item{\code{bamParams}}{An object of class ScanBamParam.
See ?Rsamtools::ScanBamParam.}

\item{\code{parallel}}{A parallel backend of the respective class. See
BiocParalell for the options}

\item{\code{processFunction}}{A custom function on how to process raw data. Not
used if center is TRUE/FALSE.}

\item{\code{optimMethod}}{The optiomisation method to be used in cross validation.}

\item{\code{optimControl}}{Settings for the optim() function.}
}}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

