% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAMDataSet-class.R
\name{makeTestGenoGAMDataSet}
\alias{makeTestGenoGAMDataSet}
\title{Make an example /code{GenoGAMDataSet}}
\usage{
makeTestGenoGAMDataSet()
}
\value{
A /code{GenoGAMDataSet} object
}
\description{
Make an example /code{GenoGAMDataSet}
}
\examples{
test <- makeTestGenoGAMDataSet()
}

