% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAM-class.R
\docType{methods}
\name{subsetByOverlaps,GenoGAM,ANY-method}
\alias{subsetByOverlaps,GenoGAM,ANY-method}
\title{Subset by overlaps method for \code{GenoGAM}}
\usage{
\S4method{subsetByOverlaps}{GenoGAM,ANY}(query, subject)
}
\arguments{
\item{query}{A \code{GenoGAM} object.}

\item{subject}{A \code{GRanges} object}

\item{...}{Additional parameters}
}
\value{
A subsetted \code{GenoGAM} object.
}
\description{
Subsetting the \code{GenoGAM} by a \code{GRanges} object
}
\examples{
gg <- makeTestGenoGAM()
gr <- GRanges("chrI", IRanges(1,40))
subsetByOverlaps(gg, gr)
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

