% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoGAMDataSet-class.R
\docType{methods}
\name{subsetByOverlaps,GenoGAMDataSet,GRanges-method}
\alias{subsetByOverlaps,GenoGAMDataSet,GRanges-method}
\title{Subset by overlaps method for \code{GenoGAMDataSet}}
\usage{
\S4method{subsetByOverlaps}{GenoGAMDataSet,GRanges}(x, ranges,
  maxgap = -1L, minoverlap = 0L, type = c("any", "start", "end", "within",
  "equal"), ...)
}
\arguments{
\item{x}{A \code{GenoGAMDataSet} object.}

\item{ranges}{A \code{GRanges} object}

\item{maxgap, minoverlap, type}{See \code{?\link[IRanges]{findOverlaps}} in
the \pkg{IRanges} package for a description of these arguments.}

\item{...}{Additional parameters}
}
\value{
A subsetted \code{GenoGAMDataSet} object.
}
\description{
Subsetting the \code{GenoGAMDataSet} by a \code{GRanges} object
}
\examples{
ggd <- makeTestGenoGAMDataSet()
gr <- GRanges("chrI", IRanges(1,50))
res <- subsetByOverlaps(ggd, gr)
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

