% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicTiles-class.R
\docType{methods}
\name{subsetByOverlaps,GenomicTiles,GRanges-method}
\alias{subsetByOverlaps,GenomicTiles,GRanges-method}
\title{Subset by overlaps method for \code{GenomciTiles}}
\usage{
\S4method{subsetByOverlaps}{GenomicTiles,GRanges}(x, ranges, maxgap = -1L,
  minoverlap = 0L, type = c("any", "start", "end", "within", "equal"), ...)
}
\arguments{
\item{x}{A \code{GenomicTiles} object.}

\item{ranges}{A \code{GRanges} object}

\item{maxgap, minoverlap, type}{See \code{?\link[IRanges]{findOverlaps}} in
the \pkg{IRanges} package for a description of these arguments.}

\item{...}{Additional parameters}
}
\value{
A subsetted \code{GenomicTiles} object.
}
\description{
Subsetting the \code{GenomicTiles} by a \code{GRanges} object
}
\examples{
gt <- makeTestGenomicTiles()
gr <- GRanges(c("chrI", "chrII"), IRanges(c(1, 120), c(40, 150)))
res <- subsetByOverlaps(gt, gr)
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

