% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakCalling.R
\name{writeToBEDFile}
\alias{writeToBEDFile}
\title{Write peaks to BED6+3/4 format}
\usage{
writeToBEDFile(peaks, file = NULL)
}
\arguments{
\item{peaks}{A data.table or data.frame of peaks as produced by callPeaks()}

\item{file}{A file name without suffix. It will be determined automatically. If no
file is given, it will be written to a generic 'peaks_[timestamp]' file in the
current working directory}
}
\value{
Nothing. A narrowPeaks or broadPeaks file written to 'file'
}
\description{
A function to write the data.table of peaks into a narrowPeaks or broadPeaks file
}
\author{
Georg Stricker \email{georg.stricker@in.tum.de}
}

